Separate RobotPrinter to its own files
