+++ lib/Communication/RobotPackets.cpp
@@ -1,5 +1,4 @@
 #include "RobotPackets.h"
-#include "IRobotIOStream.h"
 
 namespace {
     void append_uint8(uint8_t *buffer, size_t &offset, uint8_t value)
@@ -72,28 +71,4 @@ bool MotionCommandPacket::serialize(uint8_t *buffer, size_t buffer_size) const
     append_uint8(buffer, offset, 0); // CRC is reserved for RobotPrinter - it will be overwritten there
 
     return true;
-}
-
-uint8_t RobotPrinter::compute_xor_CRC(const uint8_t *buffer, size_t size)
-{
-    uint8_t CRC = 0;
-
-    if(size)
-    {
-        CRC = buffer[0];
-
-        // (size - 1) -> don't include the last byte - the CRC byte
-        for(unsigned int i = 1; i < (size - 1); i++)
-        {
-            CRC ^= buffer[i];
-        }
-    }
-
-    return CRC;
-}
-
-size_t RobotPrinter::write_serialized(uint8_t *buffer, size_t size)
-{
-    buffer[size - 1] = compute_xor_CRC(buffer, size);
-    return robot_serial.write(buffer, size);
 }
