+++ lib/Communication/RobotPackets.h
@@ -115,51 +115,4 @@ struct MotionCommandPacket {
     bool serialize(uint8_t *buffer, size_t buffer_size) const;
 };
 
-class IRobotIOStream;
-
-/**
- * Class to send out robot packets
- */
-class RobotPrinter
-{
-    IRobotIOStream &robot_serial;
-    /**
-     * Compute CRC by doing xor of all bytes together
-     * @param[in] buffer chunk of data containing packet
-     * @param[in] size size of packet in bytes
-     * @return computed CRC
-     */
-    uint8_t compute_xor_CRC(const uint8_t *buffer, size_t size);
-    /**
-     * Compute packet's CRC and send it out
-     * @param[in] buffer chunk of data containing packet
-     * @param[in] size size of packet in bytes
-     * @return number of bytes written
-     */
-    size_t write_serialized(uint8_t *buffer, size_t size);
-public:
-    /**
-     * Constructor method
-     * @param[in] serial serial interface packets are sent to
-     */
-    explicit RobotPrinter(IRobotIOStream &serial) : robot_serial(serial) {};
-
-    /**
-     * Serialize packet and send it out. CRC is computed and added to the end of serialized packet before sending.
-     * @param[in] packet packet to be sent out
-     * @return number of bytes written, 0 if serialization failed
-     */
-    template <typename PacketType>
-    size_t print(const PacketType &packet)
-    {
-        uint8_t buffer[PacketType::SERIALIZED_SIZE];
-        if(!packet.serialize(buffer, PacketType::SERIALIZED_SIZE))
-        {
-            return 0;
-        }
-
-        return write_serialized(buffer, PacketType::SERIALIZED_SIZE);
-    }
-};
-
 #endif
